feat: remove leading zeros from patient id
This also moves test files into subfolder `tests`.
