+++ src/decompose_xmls/requirements.txt
@@ -1,2 +1,3 @@
 pydantic==1.10.14
 confluent-kafka==2.3.0
+regex==2024.5.15
